composable graphql client
